+++ compose/foundation/foundation/src/uikitMain/kotlin/androidx/compose/foundation/text/input/internal/TextFieldKeyEventHandler.uikit.kt
@@ -16,49 +16,5 @@
 
 package androidx.compose.foundation.text.input.internal
 
-import androidx.compose.foundation.text.KeyCommand
-import androidx.compose.foundation.text.commonKeyMapping
-import androidx.compose.foundation.text.input.internal.selection.TextFieldSelectionState
-import androidx.compose.ui.input.key.KeyEvent
-import androidx.compose.ui.input.key.isMetaPressed
-import androidx.compose.ui.input.key.isShiftPressed
-
 /** Factory function to create a platform specific [TextFieldKeyEventHandler]. */
-internal actual fun createTextFieldKeyEventHandler(): TextFieldKeyEventHandler = object : TextFieldKeyEventHandler() {
-    override fun onKeyEvent(
-        event: KeyEvent,
-        textFieldState: TransformedTextFieldState,
-        textLayoutState: TextLayoutState,
-        textFieldSelectionState: TextFieldSelectionState,
-        clipboardKeyCommandsHandler: ClipboardKeyCommandsHandler,
-        editable: Boolean,
-        singleLine: Boolean,
-        onSubmit: () -> Unit
-    ): Boolean {
-        return when(commonKeyMapping{
-            event.isShiftPressed && event.isMetaPressed
-        }.map(event)) {
-            // iOS has its own Key Input handler for these keys:
-            KeyCommand.LEFT_CHAR,
-            KeyCommand.RIGHT_CHAR,
-            KeyCommand.UP,
-            KeyCommand.DOWN,
-            KeyCommand.SELECT_LEFT_CHAR,
-            KeyCommand.SELECT_RIGHT_CHAR,
-            KeyCommand.SELECT_UP,
-            KeyCommand.SELECT_DOWN -> return false
-            else -> {
-                super.onKeyEvent(
-                    event,
-                    textFieldState,
-                    textLayoutState,
-                    textFieldSelectionState,
-                    clipboardKeyCommandsHandler,
-                    editable,
-                    singleLine,
-                    onSubmit
-                )
-            }
-        }
-    }
-}
+internal actual fun createTextFieldKeyEventHandler(): TextFieldKeyEventHandler = createIOSTextFieldKeyEventHandler()
